+++ FitToFit/FitToFit.Model/Treninzi.cs
@@ -10,7 +10,7 @@
 
         public int MaxBrojClanova { get; set; }
 
-        public int CijenaPoTerminu { get; set; }
+        public decimal CijenaPoTerminu { get; set; }
 
         public string Trajanje { get; set; } = null!;
