+++ FitToFit/FitToFit.Services/Database/Treninzi.cs
@@ -13,7 +13,7 @@ public partial class Treninzi
 
     public int MaxBrojClanova { get; set; }
 
-    public int CijenaPoTerminu { get; set; }
+    public decimal CijenaPoTerminu { get; set; }
 
     public string Trajanje { get; set; } = null!;
